+++ docs/components/pictograms/index.tsx
@@ -15,6 +15,7 @@ import guides from './guides.svg';
 import header from './header.svg';
 import heading from './heading.svg';
 import icon from './icon.svg';
+import keywordList from './keyword-list.svg';
 import linkList from './link-list.svg';
 import mainNav from './main-nav.svg';
 import packages from './packages.svg';
@@ -22,6 +23,7 @@ import progressIndicator from './progress-indicator.svg';
 import searchBox from './search-box.svg';
 import select from './select.svg';
 import sideNav from './side-nav.svg';
+import skipLink from './skip-link.svg';
 import starter from './starter.svg';
 import text from './text.svg';
 import textInput from './text-input.svg';
@@ -45,6 +47,7 @@ export const pictograms = {
 	header,
 	heading,
 	icon,
+	'keyword-list': keywordList,
 	linkList,
 	'link-list': linkList,
 	mainNav,
@@ -55,6 +58,7 @@ export const pictograms = {
 	'search-box': searchBox,
 	select,
 	'side-nav': sideNav,
+	'skip-link': skipLink,
 	starter,
 	text,
 	textarea: textInput,