+++ docs/playroom/components.js
@@ -47,3 +47,14 @@ export {
 export { Icon } from '@ag.ds-next/icon';
 export { ProgressIndicator } from '@ag.ds-next/progress-indicator';
 export { ControlGroup, Checkbox, Radio } from '@ag.ds-next/control-input';
+export {
+	SearchBox,
+	SearchBoxInput,
+	SearchBoxButton,
+} from '@ag.ds-next/search-box';
+export { KeywordList } from '@ag.ds-next/keyword-list';
+export {
+	SkipLinks,
+	SkipLinkContainer,
+	SkipLinkItem,
+} from '@ag.ds-next/skip-link';