+++ packages/main-nav/src/MainNav.tsx
@@ -1,22 +1,32 @@
 import { NavContainer, NavContainerProps } from './NavContainer';
 import { NavList, NavListProps } from './NavList';
 
+import { findBestMatch } from './utils';
+
+export type MainNavProps = React.PropsWithChildren<{
+	variant: NavContainerProps['variant'];
+	links: NavListProps['links'];
+	secondaryLinks?: NavListProps['links'];
+	activePath?: string;
+	id?: string;
+}>;
+
 export function MainNav({
 	variant = 'darkAlt',
 	links,
 	secondaryLinks,
 	activePath,
-}: React.PropsWithChildren<{
-	variant: NavContainerProps['variant'];
-	links: NavListProps['links'];
-	secondaryLinks?: NavListProps['links'];
-	activePath?: string;
-}>) {
+	id,
+}: MainNavProps) {
+	const bestMatch = findBestMatch(
+		[...links, ...(secondaryLinks || [])],
+		activePath
+	);
 	return (
-		<NavContainer variant={variant}>
-			<NavList links={links} activePath={activePath} />
+		<NavContainer variant={variant} id={id}>
+			<NavList links={links} activePath={bestMatch} />
 			{secondaryLinks?.length ? (
-				<NavList links={secondaryLinks} activePath={activePath} />
+				<NavList links={secondaryLinks} activePath={bestMatch} />
 			) : null}
 		</NavContainer>
 	);