+++ packages/main-nav/src/NavList.tsx
@@ -11,7 +11,6 @@ export type NavListProps = {
 
 export function NavList({ links, activePath }: NavListProps) {
 	const Link = useLinkComponent();
-	const bestMatch = findBestMatch(links, activePath);
 	return (
 		<Flex
 			as="ul"
@@ -20,13 +19,15 @@ export function NavList({ links, activePath }: NavListProps) {
 			css={{
 				[tokens.mediaQuery.max.sm]: {
 					'& > li': {
-						borderTop: `1px solid ${boxPalette.border}`,
+						borderTopWidth: tokens.borderWidth.sm,
+						borderTopStyle: 'solid',
+						borderTopColor: boxPalette.border,
 					},
 				},
 			}}
 		>
 			{links.map(({ href, label, ...props }, index) => (
-				<NavItem key={index} active={href === bestMatch}>
+				<NavItem key={index} active={href === activePath}>
 					<Link href={href} {...props}>
 						{label}
 					</Link>
@@ -34,21 +35,4 @@ export function NavList({ links, activePath }: NavListProps) {
 			))}
 		</Flex>
 	);
-}
-
-function findBestMatch(links: { href: string }[], activePath?: string) {
-	if (!activePath) return '';
-	let bestMatch = '';
-
-	for (const link of links) {
-		if (link.href === activePath) return link.href;
-		if (
-			activePath?.startsWith(link.href) &&
-			link.href.length > bestMatch.length
-		) {
-			bestMatch = link.href;
-		}
-	}
-
-	return bestMatch;
 }