+++ packages/skip-link/src/SkipLinkContainer.tsx
@@ -1,0 +1,15 @@
+import React, { DetailedHTMLProps, HTMLAttributes } from 'react';
+
+export type SkipLinkContainerProps = DetailedHTMLProps<
+	HTMLAttributes<HTMLElement>,
+	HTMLElement
+>;
+
+export const SkipLinkContainer = ({
+	children,
+	...props
+}: SkipLinkContainerProps) => (
+	<nav aria-label="skip links navigation" {...props}>
+		{children}
+	</nav>
+);