+++ packages/breadcrumbs/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @ag.ds-next/breadcrumbs
 
+## 4.0.0
+
+### Patch Changes
+
+- e3f309e: Remove typographic box props from `BreadcrumbsItem`
+- Updated dependencies [9680133]
+- Updated dependencies [c7d2a66]
+- Updated dependencies [cd7e437]
+- Updated dependencies [d94cb5d]
+- Updated dependencies [b6a9bd9]
+- Updated dependencies [214cbf9]
+- Updated dependencies [9680133]
+- Updated dependencies [b6a9bd9]
+- Updated dependencies [214cbf9]
+- Updated dependencies [3e8b39b]
+  - @ag.ds-next/core@2.1.0
+  - @ag.ds-next/box@4.0.0
+  - @ag.ds-next/icon@4.0.0
+  - @ag.ds-next/text@4.0.0
+
 ## 3.0.0
 
 ### Patch Changes