+++ packages/icon/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @ag.ds-next/icon
 
+## 4.0.0
+
+### Patch Changes
+
+- b6a9bd9: Removed export of `IconExamples`
+- 214cbf9: Updated the `size` prop to be required as it has a default value.
+- b6a9bd9: Renamed component files from `index.ts`
+- Updated dependencies [9680133]
+- Updated dependencies [c7d2a66]
+- Updated dependencies [cd7e437]
+- Updated dependencies [d94cb5d]
+- Updated dependencies [9680133]
+- Updated dependencies [b6a9bd9]
+- Updated dependencies [214cbf9]
+- Updated dependencies [3e8b39b]
+  - @ag.ds-next/core@2.1.0
+  - @ag.ds-next/box@4.0.0
+  - @ag.ds-next/text@4.0.0
+
 ## 3.0.0
 
 ### Minor Changes